整合胡建飞的项目【餐桌管理模块】

+++ backend-manager/backend-manager-pojo/src/main/java/cn/yutang/backend/pojo/po/DinnerTableExample.java
@@ -262,7 +262,7 @@ public class DinnerTableExample {
             return (Criteria) this;
         }
 
-        public Criteria andTbIdIn(List<Integer> values) {
+        public Criteria andTbIdIn(List<Long> values) {
             addCriterion("tb_id in", values, "tbId");
             return (Criteria) this;
         }
